Move non-common connection files
