+++ src/server/connections/snowflake_connection.ts
@@ -25,8 +25,8 @@ import {SnowflakeConnection} from '@malloydata/db-snowflake';
 import {
   ConfigOptions,
   SnowflakeConnectionConfig,
-} from '../types/connection_manager_types';
-import {GenericConnection} from '../types/worker_message_types';
+} from '../../common/types/connection_manager_types';
+import {GenericConnection} from '../../common/types/worker_message_types';
 
 export const createSnowflakeConnection = async (
   client: GenericConnection,
